requirements.txt
Twisted==13.2.0
pyOpenSSL==0.14
pyenchant==1.6.5
passlib==1.6.2
peewee==2.2.2
Yapsy==1.10.423
PyYAML==3.11
